# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  auth.DevLoginIn:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      role:
        type: string
    required:
    - phone
    type: object
  auth.DevLoginOut:
    properties:
      access_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/auth.UserRead'
    type: object
  auth.ErrorResponse:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        description: Details is present only for VALIDATION_ERROR responses.
      message:
        example: Human-readable error description.
        type: string
    type: object
  auth.OTPRequestIn:
    properties:
      phone:
        type: string
      return_url:
        type: string
    required:
    - phone
    type: object
  auth.OTPRequestOut:
    properties:
      challenge_id:
        type: string
      expires_in:
        type: integer
    type: object
  auth.OTPVerifyIn:
    properties:
      challenge_id:
        type: string
      code:
        type: string
    required:
    - challenge_id
    - code
    type: object
  auth.OTPVerifyOut:
    properties:
      access_token:
        type: string
      is_new_user:
        type: boolean
      return_url:
        type: string
      token_type:
        type: string
    type: object
  auth.ProfileCompleteRequest:
    properties:
      email:
        maxLength: 255
        type: string
      first_name:
        maxLength: 100
        minLength: 1
        type: string
      last_name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - first_name
    - last_name
    type: object
  auth.RefreshOut:
    properties:
      access_token:
        type: string
      token_type:
        type: string
    type: object
  auth.UserRead:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/db.UserRole'
      status:
        $ref: '#/definitions/db.UserStatus'
      updated_at:
        type: string
    type: object
...
  db.UserRole:
    enum:
    - guest
    - host
    - admin
    type: string
    x-enum-varnames:
    - UserRoleGuest
    - UserRoleHost
    - UserRoleAdmin
  db.UserStatus:
    enum:
    - pending_profile
    - active
    - suspended
    - banned
    type: string
    x-enum-varnames:
    - UserStatusPendingProfile
    - UserStatusActive
    - UserStatusSuspended
    - UserStatusBanned
...
  geo.GeoSuggestItem:
    properties:
      city:
        type: string
      country:
        type: string
      house_number:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      raw_address:
        type: string
      street:
        type: string
    type: object
  geo.GeoSuggestResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/geo.GeoSuggestItem'
        type: array
    type: object
...
host: localhost:8000
info:
  contact:
    email: dev@flickey.com
    name: Flickey Dev Team
  description: Property-rental platform backend. Provides OTP-based authentication,
    multi-step listing wizard, media upload workflow, amenity catalogue, and an admin
    moderation panel.
  license:
    name: MIT
  termsOfService: http://flickey.com/terms/
  title: Flickey Go API
  version: "1.0"
paths:
...
  /auth/complete-profile:
    post:
      consumes:
      - application/json
      description: Sets first name, last name, and optional email for a newly registered
        user. Transitions status from pending_profile to active.
      parameters:
      - description: Profile fields (first_name and last_name required, email optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.ProfileCompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.UserRead'
        "400":
          description: VALIDATION_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: MISSING_TOKEN / INVALID_TOKEN / TOKEN_EXPIRED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "409":
          description: EMAIL_TAKEN – email already used by another account
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete user profile
      tags:
      - Auth
  /auth/dev-login:
    post:
      consumes:
      - application/json
      description: Creates or finds a user by phone and returns tokens without OTP
        verification. **Only available when ENVIRONMENT=dev.**
      parameters:
      - description: Phone and optional profile fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.DevLoginIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.DevLoginOut'
        "400":
          description: VALIDATION_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: NOT_FOUND – endpoint disabled in non-dev environments
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Dev-only instant login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Revokes the current refresh token and clears the HttpOnly cookie.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Log out (current session)
      tags:
      - Auth
  /auth/logout-all:
    post:
      description: Revokes all refresh tokens for the current user across all devices.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: MISSING_TOKEN / INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out all sessions
      tags:
      - Auth
  /auth/me:
    get:
      description: Returns the authenticated user's profile. Automatically promotes
        guest → host if the user already has published listings.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.UserRead'
        "401":
          description: MISSING_TOKEN / INVALID_TOKEN / TOKEN_EXPIRED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Exchanges a valid refresh_token cookie (or Authorization header
        in dev) for a new access token and rotates the refresh token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.RefreshOut'
        "401":
          description: MISSING_REFRESH_TOKEN / TOKEN_EXPIRED / INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Refresh access token
      tags:
      - Auth
  /auth/request-otp:
    post:
      consumes:
      - application/json
      description: Sends a one-time password to the provided phone number and returns
        a challenge ID used in the verify step.
      parameters:
      - description: Phone number (E.164) and optional return URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.OTPRequestIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.OTPRequestOut'
        "400":
          description: VALIDATION_ERROR – missing or invalid phone
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "429":
          description: RATE_LIMIT_EXCEEDED – too many OTP requests
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Request an OTP code
      tags:
      - Auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Validates the 6-digit OTP against the challenge ID. On success
        issues an access token in the body and sets an HttpOnly refresh_token cookie.
      parameters:
      - description: Challenge ID and 6-digit OTP code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.OTPVerifyIn'
      produces:
      - application/json
      responses:
        "200":
          description: access_token (JWT bearer) + is_new_user flag
          schema:
            $ref: '#/definitions/auth.OTPVerifyOut'
        "400":
          description: VALIDATION_ERROR – missing fields or wrong OTP format
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: INVALID_OTP / OTP_EXPIRED / CHALLENGE_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "429":
          description: RATE_LIMIT_EXCEEDED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Verify OTP and authenticate
      tags:
      - Auth
  /geo/suggest:
    get:
      description: Returns geocoded address suggestions and coordinates for a given
        search query (proxied to Photon OSM with Redis caching).
      parameters:
      - description: Search query string (min 2 characters)
        in: query
        name: q
        required: true
        type: string
      - description: 'Language code (default: ru)'
        in: query
        name: lang
        type: string
      - description: 'Max suggestions to return (default: 5, max: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/geo.GeoSuggestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Address autocomplete suggestions
      tags:
      - Geo
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: **Bearer {token}**'
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: HttpOnly refresh token cookie set by /auth/verify-otp and rotated
      by /auth/refresh.
    in: cookie
    name: refresh_token
    type: apiKey
swagger: "2.0"